update upadte-form alert

+++ public/js/book/form.js
@@ -22,9 +22,11 @@ if(document.querySelector('#btRemoveCover'))
 	document.querySelector('#btRemoveCover').addEventListener('click', onRemoveFile)
 
 function onRemoveFile(e) {
-	var idx = this.dataset['idx'];
-	var parent = this.parentNode;
-	axios.delete('/api/book/file/'+ idx).then(onSuccess).catch(onError)
+	if(confirm('파일을 삭제하시겠습니까?\n삭제하신 파일은 되돌릴 수 업습니다.')) {
+		var idx = this.dataset['idx'];
+		var parent = this.parentNode;
+		axios.delete('/api/book/file/'+ idx).then(onSuccess).catch(onError)
+	}
 	function onSuccess(r) {
 		if(r.data.code === 200) parent.remove()
 	}
